+++ poulpy-core/src/test_suite/mod.rs
@@ -53,6 +53,8 @@ macro_rules! core_backend_test_suite {
                     $crate::test_suite::noise::linear_transformation::test_glwe_hoisted_baby_rotations_match_automorphism,
                 glwe_eval_linear_transformation_skips_empty_giant_steps =>
                     $crate::test_suite::noise::linear_transformation::test_glwe_eval_linear_transformation_skips_empty_giant_steps,
+                glwe_prepared_giant_prods_match_sequential =>
+                    $crate::test_suite::noise::linear_transformation::test_glwe_prepared_giant_prods_match_sequential,
                 glwe_packing => $crate::test_suite::noise::test_glwe_packing,
                 gglwe_switching_key_encrypt_sk => $crate::test_suite::noise::encryption::test_gglwe_switching_key_encrypt_sk,
                 gglwe_switching_key_compressed_encrypt_sk =>